+++ pkg/authenticator/requests_test.go
@@ -8,14 +8,13 @@ import (
 
 func TestRequests(t *testing.T) {
 	Convey("LoginRequest", t, func() {
-		conjurVersion := "5"
 		authnURL := "dummyURL"
 		csrBytes := []byte("dummyCSRBytes")
 
 		Convey("Given a host's username prefix", func() {
 			usernamePrefix := "host.path.to.policy"
 
-			req, err := LoginRequest(authnURL, conjurVersion, csrBytes, usernamePrefix)
+			req, err := LoginRequest(authnURL, csrBytes, usernamePrefix)
 			Convey("Finishes without raising an error", func() {
 				So(err, ShouldBeNil)
 			})